requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
cairocffi==1.7.1
camel-ai==0.2.78
camel-oasis==0.2.5
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
chardet==7.1.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
emoji==2.15.0
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
identify==2.6.18
idna==3.11
igraph==0.11.6
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
langdetect==1.0.9
lazy-object-proxy==1.12.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mpmath==1.3.0
mypy_extensions==1.1.0
neo4j==5.23.0
networkx==3.6.1
nltk==3.9.3
nodeenv==1.10.0
numpy==2.4.3
oauthlib==3.3.1
openai==1.109.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
packaging==26.0
pandas==2.2.2
pathable==0.4.4
pillow==10.3.0
platformdirs==4.9.4
pluggy==1.6.0
prance==23.6.21.0
pre-commit==3.7.1
psutil==5.9.8
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
pypdf==6.9.0
pypdfium2==5.6.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.36.2
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rpds-py==0.30.0
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.0
six==1.17.0
slack_sdk==3.31.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.2
starlette==0.52.1
sympy==1.14.0
tabulate==0.10.0
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.13.7
unstructured-client==0.42.10
urllib3==2.6.3
uvicorn==0.42.0
virtualenv==21.2.0
websockets==15.0.1
wrapt==2.1.2
